fix: audit truncating casts (H10)
Client-influenced sites get checked conversions: BlockStream's range
seek/skip fails with InvalidInput when a Range header exceeds the
address space, and the read cap saturates so end == u64::MAX cannot
overflow. debug_assert size comparisons flip to the lossless direction.
Provably-bounded sites (block id widths, path length, modulus-bounded
test data) carry #[allow] with the bounding argument.
-W clippy::cast_possible_truncation is clean across all targets.

+++ s3cas/src/check.rs
@@ -200,7 +200,7 @@ async fn read_object(
     }
 
     let block_size: usize = paths.iter().map(|(_, size)| size).sum();
-    debug_assert!(obj_meta.size() as usize == block_size);
+    debug_assert!(obj_meta.size() == block_size as u64);
 
     let mut block_stream = BlockStream::new(paths, block_size, RangeRequest::All, metrics.to_cas());
     let mut data = Vec::with_capacity(block_size);
@@ -249,6 +249,7 @@ mod tests {
     }
 
     /// Stores one multi-block object and hands back the store it lives in.
+    #[allow(clippy::cast_possible_truncation)] // the modulus bounds the cast
     async fn store_object(dir: &TempDir, opts: &StoreOptions) -> CasFS {
         let casfs = casfs_at(dir, opts);
         casfs.create_bucket("bucket").unwrap();
